refactor(input): refactoring CTRL R
in order to refactor the CTRL R, the keyboard dispatcher has been restructured in a new file and refer now to both CTRL R and standard input instead of using two distinct dispatchers.
Still need to improve the dispatcher mechanism.
Still need to refactor other parts of the input code.