feat(input): refactoring input
input.c has been widely modified. The objective of this refactor was to modify the serial algorithm in order to provide a derivation algorithm. In this way, and thanks to the new independence of ctrl r and standard input,  the code is now modular and allow further implementation in the dispatcher.
if this refactor has been successfully implemented, i still need to fix some minor problem.