feat(expansions): added simple dash expansion #2
I made a mistake in the precedent commit (89a3d9). Indeed, it refers to the sharp (single begin sharp) and not to the dash.
Added now the real dash expansion management : see ${var_name:-word} = if var_name exists and not null -> expand of var_name, otherwise it expands the word.