fix(expansions): testing and fixing the expansions #2
fixed and tested the management of sharp expansions
fixed and tested the complex expansion management
fixed and tested the star pattern matching management
added the unitary tests files ad hoc
still need to test and fix the management of other pattern matching and the multi-pattern matching.
still need to managed the leaks and other specifics tests.