feat(set): added positional parameters management
In order to complete the shell variables and expansions features, i have added the positional parameters management.
you can now set a list of positional parameters with the builtin set followed buy arguments which will be referred as positional arguments.
In order to make this feature i have added a new global list which has the same characteristics as the list of shell variables.
correctly freed in the program when we exit.