fix(globing/expansions): fixed erros - see below
errors on expansion tests are fixed.
globing is fixed and now works well.
free on gnt.c has to be fixed because it can't freed the result of
globing. (cf. gnt.c).
fix on simple_expansion. works well.
still need to norm and make more test on expansions and globing.