organize(baavril's files and directories): organized and renamed files
in order to provide a better lecture of our 42sh program i organized my directories and files.
i've created a pattern matching directory and moved it in transversal directory. Indeed the pattern_matching is used both in globing and expansions/substitutions modules.
i've moved the main.c 42sh file in the transversal directory too because it the most transversal file of the program and because my eyes bleed to see only one file.c among the other src/directories.
organization is not a hard science and we can discuss those choices that reflect my interpretation of the program.