Feat(redirection): char device support
As we had talked about wednesday, I added in the support for devices
such as the most important /dev/null, which is a standard for all tests.
I still have a problem with "ls > /dev/nill" : this command executes an
ls but redirections do not apply : we do not even enter the i_less
function... See issue.